+++ pkg/scheduler/actions/allocate/allocateGpuMemory_test.go
@@ -182,5 +182,69 @@ func getMemoryGPUTestsMetadata() []integration_tests_utils.TestTopologyMetadata 
 				},
 			},
 		},
+		{
+			TestTopologyBasic: test_utils.TestTopologyBasic{
+				Name: "Pending job requests gpu memory while other job terminates",
+				Jobs: []*jobs_fake.TestJobBasic{
+					{
+						Name:                  "pending_job-0",
+						RequiredGpuMemory:     50,
+						RequiredMemoryPerTask: 1500,
+						Priority:              constants.PriorityBuildNumber,
+						QueueName:             "queue0",
+						Tasks: []*tasks_fake.TestTaskBasic{
+							{
+								State:     pod_status.Pending,
+								GPUGroups: []string{"0"},
+							},
+						},
+					},
+					{
+						Name:                  "running_job-0",
+						RequiredMemoryPerTask: 1000,
+						Priority:              constants.PriorityBuildNumber,
+						QueueName:             "queue0",
+						Tasks: []*tasks_fake.TestTaskBasic{
+							{
+								State:     pod_status.Releasing,
+								GPUGroups: []string{"0"},
+								NodeName:  "node0",
+							},
+						},
+					},
+				},
+				Nodes: map[string]nodes_fake.TestNodeBasic{
+					"node0": {
+						GPUs:      1,
+						CPUMemory: 2000,
+					},
+				},
+				Queues: []test_utils.TestQueueBasic{
+					{
+						Name:         "queue0",
+						DeservedGPUs: 1,
+					},
+				},
+				JobExpectedResults: map[string]test_utils.TestExpectedResultBasic{
+					"pending_job-0": {
+						Status:         pod_status.Pipelined,
+						MemoryRequired: 1500,
+						GPUGroups:      []string{"0"},
+					},
+					"running_job-0": {
+						Status:         pod_status.Releasing,
+						GPUGroups:      []string{"0"},
+						MemoryRequired: 1000,
+						NodeName:       "node0",
+					},
+				},
+				Mocks: &test_utils.TestMock{
+					CacheRequirements: &test_utils.CacheMocking{
+						NumberOfCacheBinds:      0,
+						NumberOfPipelineActions: 1,
+					},
+				},
+			},
+		},
 	}
 }